+++ float64/conj/lib/native.js
@@ -1,7 +1,7 @@
 /**
 * @license Apache-2.0
 *
-* Copyright (c) 2018 The Stdlib Authors.
+* Copyright (c) 2025 The Stdlib Authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -20,13 +20,16 @@
 
 // MODULES //
 
-var real = require( './../../../float64/real' );
-var imag = require( './../../../float64/imag' );
 var Complex128 = require( './../../../float64/ctor' );
+var addon = require( './../src/addon.node' );
+
+
+// MAIN //
 
 /**
 * Returns the complex conjugate of a double-precision complex floating-point number.
 *
+* @private
 * @param {Complex128} z - complex number
 * @returns {Complex128} complex conjugate
 *
@@ -39,7 +42,8 @@ var Complex128 = require( './../../../float64/ctor' );
 * // returns <Complex128>[ 5.0, -3.0 ]
 */
 function conj( z ) {
-	return new Complex128( real( z ), -imag( z ) );
+	var v = addon( z );
+	return new Complex128( v.re, v.im );
 }